+++ x/evm/keeper/grpc_query.go
@@ -422,6 +422,7 @@ func execTrace[T traceRequest](
 	c context.Context,
 	req T,
 	k Keeper,
+	baseFee *big.Int,
 	msgCb func(
 		ctx sdk.Context,
 		cfg *EVMConfig,
@@ -459,6 +460,10 @@ func execTrace[T traceRequest](
 		return nil, status.Errorf(codes.Internal, "failed to load evm config: %s", err.Error())
 	}
 
+	if baseFee != nil {
+		cfg.BaseFee = baseFee
+	}
+
 	msg, err := msgCb(ctx, cfg, req.GetTraceConfig())
 	if err != nil {
 		return nil, status.Error(codes.Internal, err.Error())
@@ -482,10 +487,15 @@ func execTrace[T traceRequest](
 // executes the given message in the provided environment. The return value will
 // be tracer dependent.
 func (k Keeper) TraceTx(c context.Context, req *types.QueryTraceTxRequest) (*types.QueryTraceTxResponse, error) {
+	var baseFee *big.Int
+	if req != nil && req.BaseFee != nil {
+		baseFee = big.NewInt(req.BaseFee.Int64())
+	}
 	resultData, err := execTrace(
 		c,
 		req,
 		k,
+		baseFee,
 		func(ctx sdk.Context, cfg *EVMConfig, traceConfig *types.TraceConfig) (*core.Message, error) {
 			signer := ethtypes.MakeSigner(cfg.ChainConfig, big.NewInt(ctx.BlockHeight()))
 			tracer, err := newTacer(&logger.Config{}, cfg.TxConfig, traceConfig)
@@ -608,6 +618,7 @@ func (k Keeper) TraceCall(c context.Context, req *types.QueryTraceCallRequest) (
 		c,
 		req,
 		k,
+		nil,
 		func(ctx sdk.Context, cfg *EVMConfig, _ *types.TraceConfig) (*core.Message, error) {
 			var args types.TransactionArgs
 			err := json.Unmarshal(req.Args, &args)